last update

+++ Src/MentalHealthcare.Application/HelpCenterItem/Commands/Delete/DeleteHelpCenterItemHandler.cs
@@ -1,4 +1,7 @@
 using MediatR;
+using MentalHealthcare.Application.SystemUsers;
+using MentalHealthcare.Domain.Constants;
+using MentalHealthcare.Domain.Exceptions;
 using MentalHealthcare.Domain.Repositories;
 using Microsoft.Extensions.Logging;
 
@@ -6,13 +9,18 @@ namespace MentalHealthcare.Application.HelpCenterItem.Commands.Delete;
 
 public class DeleteHelpCenterItemHandler(
     ILogger<DeleteHelpCenterItemHandler> logger,
-    IHelpCenterRepository helpCenterRepository
+    IHelpCenterRepository helpCenterRepository,
+    IUserContext userContext
 ) : IRequestHandler<DeleteHelpCenterItemCommand>
 {
     public async Task Handle(DeleteHelpCenterItemCommand request, CancellationToken cancellationToken)
     {
-        //todo
-        //add auth
+        var currentUser = userContext.GetCurrentUser();
+        if (currentUser == null || !currentUser.IsAuthorized([UserRoles.Admin]))
+        {
+            logger.LogWarning("Unauthorized attempt to delete HelpCenter Item by user: {UserId}", currentUser?.Id);
+            throw new ForBidenException("Don't have the permission delete HelpCenter Item");
+        }
         logger.LogInformation(@"Delete Term with id {}", request.HelpCenterId);
         await helpCenterRepository.DeleteAsync(request.HelpCenterId);
         logger.LogInformation(@"Term with id {} deleted", request.HelpCenterId);
